Add script to run all types of tests
Note this won't run feature tests if any unit tests fail. That's
intentional, to focus on fixing the failures before running the full
suite. Also intentional that the unit tests run first, as they tend to
be quicker to execute, so we want to fail fast if we do fail.

If this app were to grow and eventually run in production, a future
strategy would be to run tests in parallel on a CI environment.