Configure Prettier
With personal preferences. Note that SvelteKit is not using all Prettier
defaults, it overrode a few options. It's an opinionated framework.

I was going to switch from tabs back to the default spaces but learned
today that tabs are more accessible for braille readers and will
likely become the default in Prettier v3 [1].

[1]: https://github.com/prettier/prettier/issues/7475