Add List/Item types
Declare new internal (shared) types/interfaces in one place. If the
number of types grows significantly, we can change the types file
to a types directory.

Likewise with the test data function in `database.test.ts`, we can
extract that to a `mockData.ts` file later for example if it helps.

P.S. I don't normally write out as detailed a thought process as this in
my commit messages. Only to help the reviewer(s) know my thought process.