Add a /lists page and basic browser test
We could also have put the lists straight onto the index page, a la
Google Keep, but decided to keep it more conventional and have the route
follow the resource, and this way the landing page can be used for
marketing, sign up, etc. With the idea being once you sign in, you get
redirected to the lists page. (Theoretical, out of scope for demo.)