Test loaded list titles show on page
Had to cheat a bit because the test data with the setup that was working
around Playwright's import restrictions wasn't working, data wasn't
being stored. Very briefly looked into before changing strategy to
get the test to pass and then improve it later.

Created a database seed that only runs in non-production environments.

We can get rid of the seed once more functionality exists and we
extend the end-to-end test to also create the list (via browser
automation, not test helpers) before asserting their existence. That way
we also test more of the high-level functionality.

Having the seed is nice for demo/dev purposes though, as you don't start
with a blank screen, but it's not code that should live among production
code. If we had a persistent store, we could move it to a dev script.

Note there is a lot of Svelte/SvelteKit-isms happening in this commit,
which will be very confusing if unfamiliar with the libraries. For
example, SvelteKit will ignore any files under `routes/` that do not
have a `+` prefix. The `+` pages with `server` in the name run on the
backend. The server file, which is specific to the accompanying page
file alongside it, contains a `load` function which is a pattern
SvelteKit uses to load page data (e.g. from the database), and the
return value becomes available on the page as `PageData`. In the
`.svelte` pages, we have to `export` properties to make them act as
component "props" or access them in the page. We're also using Svelte
syntax on the page to iterate through the page data and dynamically make
a list of titles. By the way, it is possible to pass data between client
and server in more traditional ways (e.g. endpoints + fetch) but
SvelteKit is an opinionated framework recommends you do it it their way.
Using form actions is another way to pass data, which I'll be using.