Change expected behaviour for getList sad path
It's not exceptional behaviour when a query does not return a matching
record. Instead we can have the caller, i.e. the controller action,
return an error when a list you expect to exist because some reference
to it exists, is missing, that's more exceptional behaviour.

Note the distinction between returning undefined, void, or null.
Void is also undefined but it's specified at the return type level in TS
and would imply we don't care about the return value, typically because
we only care about some side effect that happened. But here we do care
because the return lets us differentiate between a found/missing item.
As for undefined vs null, in some languages undefined doesn't exist and
so we can only use null, but here we can use it as a distinction. Null
can imply that the item does exist but it doesn't point to a value (i.e.
it is intentional). Whereas undefined tells us it doesn't exist at all.