Can edit and save a list (title)
Note the `save` action ends by calling an async function `db.saveList`
without using `await`. This is intentional. If we don't call `await`
then the called function runs asynchronously in the background. We want
that behaviour when editing a list because the save should happen in the
background without being blocking, we're not relying on the saved data
to continue using the list, if we were we would need to wait first.

Also moved the List frontend code to its own component to encapsulate it
better. Also makes it easier to reuse if needed later.