Error on saving an edited list that doesn't exist
We need to fetch the persisted list to update it with the changes.
If the list cannot be found, we can't update it, hence an error.

Instead of throwing an error, which in this case redirects them to a 404
page, we could have had the controller action return an object response
such as `{ list, success: false }` or the more elaborate and extendable
form of `{ list, errors: [{ message: 'List not found' }] }`. Then the
page can check if it was successful and take relevant steps such as
displaying an error on the same page, perhaps in a flash/toast message.

This could work for the '404 List not found' but then we'd need to give
the page extra logic to handle a redirect because it will be a repeated
error if they try saving again on the same page. I find it cleaner to
handle that logic in the backend.

However if the actual `db.saveList()` operation were to fail, this would
be a good place to return an object response for the caller to decide
how to handle the error because in a real app we can have various errors
at that stage. In this toy app, the `db.saveList` operation will never
fail in its current state, it's super simple and saves in-memory.

I'm also not testing this error flow, don't think it's worth it, too
little value for this toy app. In a prod app we could test the missing
list in a unit test and also the db error on save with a mock.