Don't throw an error and redirect on load failure
Was a good start as it delayed extra code but no longer sufficient. We
now want the client to have control over how to handle errors.

This has become important now because we want to start differentiating
on error behaviours. For example, if we delete a list and the page
auto-reloads, we don't want to get a 404 error because the list we just
deleted failed to load, that puts the user in a dead-end. The client can
instead decide whether they navigate elsewhere, show the error, hide the
error, and so on, rather than having the server dictate.

If you're still reading my commit messages and wondering why I'm up
late, it's because I was watching the Superbowl. Good game, but my sleep
schedule is going to be messed up now :(