Progressively enhance form on /list/:id page
Form actions in SvelteKit works without client-side JavaScript by
default. But when JavaScript is available we can progressively enhance
form interactions to provide a better user experience.

Adding `use:enhance` to a form will emulate browser-native behaviour,
without the full-page reloads.

For example
- updates the `form` property on response (if action is on same page)
- resets the `form` element and invalidates all data on success
- calls `goto` on a redirect response
- ...and more

We want it so that on form submission we don't have the URL change to
include the form action path. For example, without JS, the URL would
change to `/list/:id?/save` on saving a list, now it stays `/lists/:id`.