Undo usage of progressively enhanced form
It had an unexpected side effect of resetting the ActionData which we
assign to the exported `form` variable in `+page.svelte`. Meaning the
client can no longer see the response from the actions on the server.

Not yet sure if this is a bug in SvelteKit or if I am using it
incorrectly. In the mean time I'm reverting because we need the client
to be able to see the server response. Have opened a discussion on GH.

Reverts 5ee7cd6a197b285d163ea0e2b5fc18fc7745cdef