Simplify response from server list actions
Typically when dealing with errors my go to is to use a list of objects
and it's very extendable and works for the single and multi case. But
this is a toy app and we're able to keep the errors simple. So I decided
to use a simpler approach and can reduce the client code a little.

Note that errors in `data` are from page load, while errors in `form`
are from form submission (i.e. calling an action).