Start extracting feature tests
Will remove the overlap from the old test once the new test is green.

The single feature test was getting very large making it difficult to
specify the expected behaviour in the test description, making it
difficult to follow what the test is doing as it has many
responsibilities, and slows down test execution as we're running
a single long test rather than multiple smaller tests in parallel.

There has to be a balance. You don't want your end-to-end browser tests
to be as small as unit tests because they're expensive to run. So you
want your tests to be longish and group related assertions together, but
try and keep it focused to a single user flow.