Playwright tests in same file run parallel
To increase execution speed.

Note this will break some of the existing tests as there will be state
interference due to having a global store. Can be fixed in a couple ways

- Data is stored more isolated, for example sectioned by user or session

- Make the data store injectable and use a separate instance per test

- Change the assertions to work to be specific to the data created by
  its own test (for example, instead of counting all lists which is
  subject to interference, assert that the page contains the list name)

The last choice is my preference, will go with that for now.

Note fully cleaning the database before each test is not a feasible
option, could break tests that are currently running in parallel.