Can add new items to a list
This version uses the 'New item' input field to add entered text as
a new item to the list when the 'Save list' button is pressed. Note it
currently only saves the text description, not yet the check status.

More functionality to be added in the next iterations. That includes
saving the new item checked status, and being able to add multiple items
to the list before clicking the 'Save list' button.