Changes to list items are saved
The form data basically comes in as a 2D array, for example

```
formData = [
  [ 'title', 'test title' ],
  [ 'item-count', '3' ],
  [ 'items[0]', 'on' ],
  [ 'items[0]', 'one' ],
  [ 'items[1]', 'on' ],
  [ 'items[1]', 'two' ],
  [ 'items[2]', 'three' ],
  [ 'new-item-text', 'four' ]
]
```

The item inputs (checkbox and textbox) use the name `items[i]` meaning
the submitted form data will use that as the item's values key.

Because we don't specify a `value` on the list item checkboxes, the
default value 'on' is used when selected.

In the above snippet, items "one" and "two" are checked.
Item "three" is not. And a new item "four" is being added unchecked.

Starting adding comment sections to the test as it's getting long and
difficult to read. Should break it up into multiple smaller tests soon.