Can remove an item from a list
I considered adding an id to the list items and then deleting by
filtering items in the list that do not match the deletion id. But
there's no strong need for an item id yet, we can keep deferring it.

On List.svelte L46 we have `items = items` which seems redundant.
We need this to trigger a Svelte cycle. Svelte's reactivity is only
triggered by assignment, so when we do a mutation like `splice`, we need
to force an assignment. Without that, Svelte wouldn't have detected that
`items` changed and wouldn't re-render the page to show the removed item.
See https://svelte.dev/tutorial/updating-arrays-and-objects for more.

Also note that some `data-testid` values are not unique. It's convention
to have them be unique, similar to `id` attributes. But this was
intentional for the repeated inputs, because I prefer using the
Playwright API `first()`, `last()`, `nth()` to choose an input from
a group instead of having to define new `const`s for every input locator
because their data ids are unique.