Add hidden metadata to list form
This removes the warnings that properties were being passed to the List
component but not being used. We now include them as hidden metadata.

The ID and timestamps are all visually hidden. Timestamps might be
interesting to expose to the user in the future but it's not critical.

ID is also hidden from assistive technology (e.g. screen readers)
because a long UUID will just be noise for the user and it is also in
the page's URL if they want to know it.

Because the fields are disabled, they are excluded from form submission.