NodeJS LTS 18.14.0 is preferred runtime
Note the presence of this file doesn't enforce the runtime version. It's
a recommendation, and some tools will auto-respect it and switch to that
version if you have it installed, or make it easy to install if missing.

We could be stricter by adding "engines" for Node and NPM to the
`package.json` but as this is not a production project, I won't bother
adding that as an extra barrier.

Why v18? Long term support and also likely to be installed on machines.