Create an item schema
To be used for validation. Note this level of unit testing is a bit much
especially given the schema body definition is two lines and easy to
read and understand. I added this many tests to start with because the
behaviour of the Zod library was unfamiliar to me and the documentation
and examples weren't always clear. In a prod app I would add less
coverage and focus more on happy paths for schema tests, unless there's
a reason to be extra defensive depending on context.