New item auto-adds on typing
Like how Google Keep does it. When you start typing in the new item
text field then a new item appears at the bottom of the list and it
auto-focuses so you can continue typing in it.

This change means we can add multiple items to the list at once before
saving, compared to the previous behaviour where we were limited to one
item at a time as we had to save the list for the new item to be added.