Don't require a non-empty string for list items
Was a bit too strict in hindsight. Could see it being confusing or
undesired behaviour. Note we still require a title to be non-empty,
at least while we're only showing titles on the /lists page. Once we
change that page to show list cards instead, can be less strict with
titles as well and allow it xor the body to be empty, like Google Keep.