Improve list modification
Create an intermediary schema, formListSchema, which handles the form
data from submission and then passes the output as input to listSchema
to create the validated list with modifications.

The first schema, formListSchema, verifies that the form data has
everything that's required to save a list. It then returns an object
that looks similar to a `List` type but doesn't contain all the data,
which we call `FormList`. Its data also hasn't been validated yet to
comply with a `List`. The next schema, `listSchema`, validates the
`FormList` data and converts it to a `List` if possible.

Also improve the error handling on the backend.