Add Tailwind CSS + Plugins
I've never used Tailwind before so I wouldn't be so quick to add it to
a client/production project over something I have used. This challenge
was a good learning opportunity to try it; I've heard a lot of praise.

Plugins
- tailwindcss-forms [1]
  - form default styles that are easy to override/extend
- tailwindcss-typography [2]
  - typographic default styles
- daisyui [3]
  - a popular component library for tailwind

Tailwind and the plugins were installed and auto-configured with
https://github.com/svelte-add/tailwindcss.

`npx svelte-add@latest tailwindcss --daisyui --typography --forms`

1: https://github.com/tailwindlabs/tailwindcss-forms
2: https://github.com/tailwindlabs/tailwindcss-typography
3: https://github.com/saadeghi/daisyui