Update landing page and redirect to lists
Normally this would be the page where you do stuff like marketing and
ask the user to register or log in, or redirect to the user content if
a session is present. In this case it's faked. The button does do a POST
request to the backend but we're not authenticating or storing a cookie.
We also don't support multiple users, we're not using an external store.