Hide accessibility titles
We can't use `display: none` to hide accessibility elements, that is,
content that we want to be available to accessibility tools like screen
readers and not just available but in the right place in the page.
Because `display: none` would remove the content from the page
completely and therefore it wouldn't appear to tooling. A common
solution is to use hidden classes, a variety of these exist. A good
explanation is provided on Nomensa [1].

1: https://www.nomensa.com/blog/how-improve-web-accessibility-hiding-elements