Improve delete list flow
When a list was deleted with the related form action, the page and its
components would be re-rendered. But the URL would still have the
'?/delete' appended meaning the delete form action is called again,
though this time the error handling is executed and the page prints that
the list is not found.

The newer error handling in the delete form action broke the client side
redirect we added in `+page.svelte` for going to the `/lists` page after
a list had been deleted.

But client side redirect never worked well to begin with. The execution
is too slow, and you can see SvelteKit loading the list page after list
deletion before redirecting. It's not the seamless experience we want.

So I'm switching to redirecting on the server side right after the list
is deleted. This fixes the redirect to work again and makes it seamless.